exception_fini

+++ src/main/java/com/example/UserManager/service/UserService.java
@@ -1,24 +1,21 @@
 package com.example.UserManager.service;
 
-
+import com.example.UserManager.entity.Role;
 import com.example.UserManager.entity.User;
+import com.example.UserManager.exception.UserException;
 import com.example.UserManager.repository.RoleRepository;
 import com.example.UserManager.repository.UserRepository;
-import org.springframework.stereotype.Service;
-import java.util.List;
-import com.example.UserManager.entity.Role;
 import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
 
+import java.util.List;
 
 @Service
 @RequiredArgsConstructor
-
 public class UserService {
 
     private final UserRepository userRepository;
     private final RoleRepository roleRepository;
-
-
 
     // get all users
     public List<User> getAllUsers() {
@@ -27,20 +24,21 @@ public class UserService {
 
     // get user by id
     public User getUserById(Integer id) {
-        return userRepository.findById(id).orElse(null);
+        return userRepository.findById(id)
+                .orElseThrow(() -> new UserException("User not found"));
     }
 
     // create new user
     public User createUser(User user) {
 
         if (userRepository.existsByEmail(user.getEmail())) {
-            throw new RuntimeException("Email already exists");
+            throw new UserException("Email already exists");
         }
 
         Role role = roleRepository.findByName("USER").orElse(null);
 
         if (role == null) {
-            throw new RuntimeException("Default role USER not found");
+            throw new UserException("Default role USER not found");
         }
 
         user.setRole(role);
@@ -50,6 +48,10 @@ public class UserService {
 
     // delete user by id
     public void deleteUser(Integer id) {
+        if (!userRepository.existsById(id)) {
+            throw new UserException("User not found");
+        }
+
         userRepository.deleteById(id);
     }
 
@@ -58,13 +60,12 @@ public class UserService {
         User user = userRepository.findById(id).orElse(null);
 
         if (user == null) {
-            throw new RuntimeException("User not found");
+            throw new UserException("User not found");
         }
 
-        // vérifier si l'email appartient déjà à un autre utilisateur
         if (!user.getEmail().equals(userDetails.getEmail()) &&
                 userRepository.existsByEmail(userDetails.getEmail())) {
-            throw new RuntimeException("Email already exists");
+            throw new UserException("Email already exists");
         }
 
         user.setName(userDetails.getName());
@@ -73,7 +74,4 @@ public class UserService {
 
         return userRepository.save(user);
     }
-
-
-
 }
